API/requirements.txt
annotated-types   
anyio             
bcrypt            
click             
colorama          
fastapi          
h11              
idna              
passlib           
pip               
psycopg2-binary   
pydantic          
pydantic_core     
sniffio           
starlette         
typing_extensions 
uvicorn           